+++ pom.xml
@@ -275,6 +275,17 @@
           <scope>test</scope>
       </dependency>
 
+      <!-- dacp service broker-->
+      <dependency>
+          <groupId>com.asiainfo</groupId>
+          <artifactId>com.asiainfo</artifactId>
+          <version>1.0</version>
+      </dependency>
+      <dependency>
+          <groupId>org.springframework</groupId>
+          <artifactId>spring-context</artifactId>
+          <version>3.1.1.RELEASE</version>
+      </dependency>
 
   </dependencies>
   <build>
