+++ src/com/asiainfo/ocmanager/rest/resource/KerberosResource.java
@@ -1,7 +1,7 @@
 package com.asiainfo.ocmanager.rest.resource;
 
 import com.asiainfo.ocmanager.rest.resource.downloadUtils.DownloadResult;
-import com.asiainfo.ocmanager.rest.resource.downloadUtils.GetFile;
+import com.asiainfo.ocmanager.rest.resource.downloadUtils.GetKeytabFile;
 
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
@@ -23,7 +23,7 @@ public class KerberosResource {
     @Path("getkeytab/{tenantId}/{username}")
     @Produces(MediaType.APPLICATION_JSON)
     public Response getKeytab(@PathParam("tenantId")String tenantId,@PathParam("username")String username){
-        Map map = GetFile.getFile(tenantId,username);
+        Map map = GetKeytabFile.returnResult(tenantId,username);
         DownloadResult dr = (DownloadResult)map.get("res");
         Response.ResponseBuilder responseBuilder = Response.ok();
         responseBuilder.type("application/json");
@@ -35,11 +35,12 @@ public class KerberosResource {
     @Path("getFile/{tenantId}/{username}")
     @Produces(MediaType.APPLICATION_OCTET_STREAM)
     public Response getFile(@PathParam("tenantId")String tenantId,@PathParam("username")String username){
-        Map map = GetFile.getFile(tenantId,username);
-        Response.ResponseBuilder responseBuilder = Response.ok(map.get("file"));
+        Map map = GetKeytabFile.returnResult(tenantId,username);
+        File file = (File)map.get("file");
+        Response.ResponseBuilder responseBuilder = Response.ok(file);
         responseBuilder.type("applicatoin/octet-stream");
-        responseBuilder.header("Content-Disposition", "attachment; filename="+((File)map.get("file")).getName());
-        responseBuilder.header("Content-Length", Long.toString(((File)map.get("file")).length()));
+        responseBuilder.header("Content-Disposition", "attachment; filename="+(file.getName()));
+        responseBuilder.header("Content-Length", Long.toString(file.length()));
         Response response = responseBuilder.build();
         return response;
     }
